WIP: refactor array utils.
