+++ phy/electrode/mea.py
@@ -9,7 +9,7 @@
 
 import numpy as np
 
-from ..utils.array import _as_array
+from ..utils._types import _as_array
 
 
 #------------------------------------------------------------------------------
