+++ phy/io/sparse.py
@@ -8,7 +8,7 @@
 
 import numpy as np
 
-from ..utils.array import _as_array
+from ..utils._types import _as_array
 
 
 #------------------------------------------------------------------------------
