+++ phy/plot/features.py
@@ -20,7 +20,8 @@ from ._vispy_utils import (BaseSpikeVisual,
                            )
 from ._panzoom import PanZoomGrid
 from ..ext.six import string_types
-from ..utils.array import _as_array, _index_of
+from ..utils._types import _as_array
+from ..utils.array import _index_of
 from ..utils.logging import debug
