+++ phy/utils/_types.py
@@ -16,6 +16,12 @@ from ..ext.six import string_types, integer_types
 # Various Python utility functions
 #------------------------------------------------------------------------------
 
+_ACCEPTED_ARRAY_DTYPES = (np.float, np.float32, np.float64,
+                          np.int, np.int8, np.int16, np.uint8, np.uint16,
+                          np.int32, np.int64, np.uint32, np.uint64,
+                          np.bool)
+
+
 class Bunch(dict):
     """A dict with additional dot syntax."""
     def __init__(self, *args, **kwargs):
@@ -49,3 +55,37 @@ def _as_list(obj):
         return [obj]
     else:
         return obj
+
+
+def _is_array_like(arr):
+    return isinstance(arr, (list, np.ndarray))
+
+
+def _as_array(arr, dtype=None):
+    """Convert an object to a numerical NumPy array.
+
+    Avoid a copy if possible.
+
+    """
+    if isinstance(arr, integer_types + (float,)):
+        arr = [arr]
+    out = np.asarray(arr)
+    if dtype is not None:
+        if out.dtype != dtype:
+            out = out.astype(dtype)
+    if out.dtype not in _ACCEPTED_ARRAY_DTYPES:
+        raise ValueError("'arr' seems to have an invalid dtype: "
+                         "{0:s}".format(str(out.dtype)))
+    return out
+
+
+def _as_tuple(item):
+    """Ensure an item is a tuple."""
+    if item is None:
+        return None
+    # elif hasattr(item, '__len__'):
+    #     return tuple(item)
+    elif not isinstance(item, tuple):
+        return (item,)
+    else:
+        return item
