+++ phy/utils/array.py
@@ -11,7 +11,7 @@ from math import floor
 import numpy as np
 
 from .logging import warn
-from ..ext import six
+from ._types import _as_tuple
 
 
 #------------------------------------------------------------------------------
@@ -110,46 +110,6 @@ def _index_of(arr, lookup):
     tmp = np.zeros(m, dtype=np.int)
     tmp[lookup] = np.arange(len(lookup))
     return tmp[arr]
-
-
-def _is_array_like(arr):
-    return isinstance(arr, (list, np.ndarray))
-
-
-_ACCEPTED_ARRAY_DTYPES = (np.float, np.float32, np.float64,
-                          np.int, np.int8, np.int16, np.uint8, np.uint16,
-                          np.int32, np.int64, np.uint32, np.uint64,
-                          np.bool)
-
-
-def _as_array(arr, dtype=None):
-    """Convert an object to a numerical NumPy array.
-
-    Avoid a copy if possible.
-
-    """
-    if isinstance(arr, six.integer_types + (float,)):
-        arr = [arr]
-    out = np.asarray(arr)
-    if dtype is not None:
-        if out.dtype != dtype:
-            out = out.astype(dtype)
-    if out.dtype not in _ACCEPTED_ARRAY_DTYPES:
-        raise ValueError("'arr' seems to have an invalid dtype: "
-                         "{0:s}".format(str(out.dtype)))
-    return out
-
-
-def _as_tuple(item):
-    """Ensure an item is a tuple."""
-    if item is None:
-        return None
-    # elif hasattr(item, '__len__'):
-    #     return tuple(item)
-    elif not isinstance(item, tuple):
-        return (item,)
-    else:
-        return item
 
 
 def _partial_shape(shape, trailing_index):
