+++ phy/waveform/filter.py
@@ -8,7 +8,7 @@
 
 from scipy import signal
 
-from ..utils.array import _as_array
+from ..utils._types import _as_array
 
 
 #------------------------------------------------------------------------------
