+++ phy/waveform/loader.py
@@ -8,7 +8,8 @@
 
 import numpy as np
 
-from ..utils.array import _as_array, _pad
+from ..utils._types import _as_array
+from ..utils.array import _pad
 from ..utils.logging import warn
